refactor repo package
