+++ cmd/helm/dependency_update_test.go
@@ -88,7 +88,7 @@ func TestDependencyUpdateCmd(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	i, err := repo.LoadIndexFile(duc.helmhome.CacheIndex("test"))
+	i, err := repo.NewChartRepositoryIndexFromFile(duc.helmhome.CacheIndex("test"))
 	if err != nil {
 		t.Fatal(err)
 	}
